# Parallelize first and last in the tsm1 engine