Also prevent the first and last optimization from being used if the
outer dimensions do not match the inner dimensions. This is because the
optimization will not return the correct points if these groupings
differ from each other.

- [ ] Rebased/mergable
- [ ] Tests pass
- [ ] CHANGELOG.md updated