@jsternberg 2017-08-29
Here are some numbers from the included benchmark first and last for TSI1. It's not as much faster as I thought it would be (and at least one of the benchmarks I ran locally it was slower, but the IDE I've been trying out may have been using too many resources from my machine at the time).

```
jsternberg:[~/g/s/g/i/influxdb] $ (js-parallelize-first-last $<>) benchcmp old.txt new.txt
benchmark                                       old ns/op     new ns/op     delta
BenchmarkEngine_CreateIterator_First_1K-8       25782         24088         -6.57%
BenchmarkEngine_CreateIterator_First_100K-8     61091         58904         -3.58%
BenchmarkEngine_CreateIterator_First_1M-8       78447         76836         -2.05%
BenchmarkEngine_CreateIterator_Last_1K-8        26488         24487         -7.55%
BenchmarkEngine_CreateIterator_Last_100K-8      68722         73596         +7.09%
BenchmarkEngine_CreateIterator_Last_1M-8        100104        97193         -2.91%

benchmark                                       old allocs     new allocs     delta
BenchmarkEngine_CreateIterator_First_1K-8       56             48             -14.29%
BenchmarkEngine_CreateIterator_First_100K-8     69             61             -11.59%
BenchmarkEngine_CreateIterator_First_1M-8       162            154            -4.94%
BenchmarkEngine_CreateIterator_Last_1K-8        56             48             -14.29%
BenchmarkEngine_CreateIterator_Last_100K-8      69             61             -11.59%
BenchmarkEngine_CreateIterator_Last_1M-8        162            154            -4.94%

benchmark                                       old bytes     new bytes     delta
BenchmarkEngine_CreateIterator_First_1K-8       9416          8632          -8.33%
BenchmarkEngine_CreateIterator_First_100K-8     25112         24328         -3.12%
BenchmarkEngine_CreateIterator_First_1M-8       35544         34760         -2.21%
BenchmarkEngine_CreateIterator_Last_1K-8        9416          8632          -8.33%
BenchmarkEngine_CreateIterator_Last_100K-8      25112         24328         -3.12%
BenchmarkEngine_CreateIterator_Last_1M-8        35544         34760         -2.21%
```