@jsternberg 2017-08-29
Upon trying to test this with `influx_stress`, I'm just not seeing any noticeable improvement. I tried a high cardinality of 450k series and a lower cardinality of 50 with having 100k points in each series and there wasn't a noticeable difference between the two. I would guess it's because there really just isn't that much time spent reading one point that parallelization really helps much.

I'll keep this PR up, but I see little reason to merge it at the moment.